LayoutTests/imported/w3c/web-platform-tests/navigation-api/ordering-and-transition/location-href-intercept-reject_no-currententrychange-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error: boo

Harness Error (FAIL), message = Unhandled rejection: boo

FAIL event and promise ordering for location.href setter intercepted by passing a rejected promise to intercept() assert_array_equals: expected property 0 to be "navigate" but got "promise microtask" (expected array ["navigate", "handler run", "promise microtask", "navigateerror", "transition.finished rejected"] got ["promise microtask", "navigate", "handler run", "navigateerror", "transition.finished rejected"])
